Enhanced explosive bullet

+++ src/main/java/com/bham/bc/components/armory/ExplosiveBullet.java
@@ -1,22 +1,13 @@
 package com.bham.bc.components.armory;
 
 
-import com.bham.bc.components.characters.GameCharacter;
 import com.bham.bc.components.characters.SIDE;
-import com.bham.bc.components.environment.GameMap;
-import com.bham.bc.components.environment.GenericObstacle;
-import com.bham.bc.components.environment.navigation.ItemType;
-import com.bham.bc.entity.BaseGameEntity;
-import com.bham.bc.entity.physics.BombTank;
-import com.bham.bc.entity.triggers.Trigger;
-import com.bham.bc.entity.triggers.TriggerRegionRectangle;
 import javafx.geometry.Point2D;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 import javafx.scene.shape.Rectangle;
 import javafx.scene.transform.Rotate;
 
-import static com.bham.bc.components.CenterController.backendServices;
 import static com.bham.bc.entity.EntityManager.entityManager;
 import static com.bham.bc.utils.Constants.FRAME_RATE;
 
@@ -26,11 +17,14 @@ public class ExplosiveBullet extends Bullet {
     public static final int WIDTH = 6;
     public static final int HEIGHT = 12;
 
-    public static final int BOUNDWIDTH = 300;
-    public static final int BOUNDHEIGHT = 300;
+    public static final int MAX_BOUNDWIDTH = 300;
+    public static final int MAX_BOUNDHEIGHT = 300;
 
-    public static final double SPEED = 5;
     public static final double DAMAGE = 100;
+
+    public static final int LIFETIME = 1*FRAME_RATE;
+
+    private int existTime;
 
     /**
      * Constructs a bullet using default bullet's attributes for speed and damage
@@ -43,12 +37,15 @@ public class ExplosiveBullet extends Bullet {
     public ExplosiveBullet(double x, double y, double angle, SIDE side) {
         super(x, y, 0, angle, side, DAMAGE);
         entityImages = new Image[] { new Image(IMAGE_PATH, WIDTH, HEIGHT, false, false) };
+        existTime = 0;
     }
 
     @Override
     public void destroy() {
-        entityManager.removeEntity(this);
-        exists = false;
+        if (existTime >= LIFETIME) {
+            entityManager.removeEntity(this);
+            exists = false;
+        }
     }
 
     @Override
@@ -63,7 +60,9 @@ public class ExplosiveBullet extends Bullet {
     }
 
     @Override
-    public void update() { move(); }
+    public void update() {
+        existTime++;
+    }
 
     @Override
     public void render(GraphicsContext gc) { drawRotatedImage(gc, entityImages[0], angle); }
@@ -71,12 +70,13 @@ public class ExplosiveBullet extends Bullet {
     @Override
     public Rectangle getHitBox() {
         //get bound rectangle top left pos
-        Point2D regionRadius = new Point2D(BOUNDWIDTH,BOUNDHEIGHT);
+        //calculate region radius according to current time
+        Point2D regionRadius = new Point2D(MAX_BOUNDWIDTH, MAX_BOUNDHEIGHT).multiply(existTime/(double)LIFETIME);
         Point2D topLeft = getPosition().subtract(regionRadius.subtract(getRadius()).multiply(0.5));
 
         //create hit box according to bound rectangle
-        Rectangle hitBox = new Rectangle(topLeft.getX(), topLeft.getY(), BOUNDWIDTH, BOUNDHEIGHT);
-        hitBox.getTransforms().add(new Rotate(angle, topLeft.getX() + BOUNDWIDTH/2,topLeft.getY() + BOUNDHEIGHT/2));
+        Rectangle hitBox = new Rectangle(topLeft.getX(), topLeft.getY(), regionRadius.getX(), regionRadius.getY());
+        hitBox.getTransforms().add(new Rotate(angle, topLeft.getX() + regionRadius.getX() /2,topLeft.getY() + regionRadius.getY() /2));
 
         return hitBox;
     }
